+++ slang/lib/src/builder/builder/generate_config_builder.dart
@@ -32,6 +32,8 @@ class GenerateConfigBuilder {
       contexts: contexts,
       interface: interfaces,
       obfuscation: config.obfuscation,
+      format: config.format,
+      autodoc: config.autodoc,
       imports: config.imports,
     );
   }