+++ slang/lib/src/builder/generator/generate_translations.dart
@@ -1,6 +1,9 @@
 import 'dart:collection';
+import 'dart:math';
 
+import 'package:collection/collection.dart';
 import 'package:slang/src/builder/generator/helper.dart';
+import 'package:slang/src/builder/model/autodoc_config.dart';
 import 'package:slang/src/builder/model/enums.dart';
 import 'package:slang/src/builder/model/generate_config.dart';
 import 'package:slang/src/builder/model/i18n_data.dart';
@@ -22,7 +25,8 @@ class ClassTask {
 /// generates all classes of one locale
 /// all non-default locales has a postfix of their locale code
 /// e.g. Strings, StringsDe, StringsFr
-String generateTranslations(GenerateConfig config, I18nData localeData) {
+String generateTranslations(GenerateConfig config, I18nData localeData,
+    List<I18nData> allTranslations) {
   final queue = Queue<ClassTask>();
   final buffer = StringBuffer();
 
@@ -32,6 +36,7 @@ String generateTranslations(GenerateConfig config, I18nData localeData) {
 ///
 // coverage:ignore-file
 // ignore_for_file: type=lint, unused_import
+${!config.format.enabled ? '// dart format off' : ''}
 ''');
 
   if (localeData.base) {
@@ -73,6 +78,7 @@ String generateTranslations(GenerateConfig config, I18nData localeData) {
       task.className,
       task.node,
       root,
+      allTranslations,
     );
 
     root = false;
@@ -96,6 +102,7 @@ void _generateClass(
   String className,
   ObjectNode node,
   bool root,
+  List<I18nData> allTranslations,
 ) {
   buffer.writeln();
 
@@ -316,10 +323,51 @@ void _generateClass(
   bool prevHasComment = false;
   node.entries.forEach((key, value) {
     // comment handling
-    if (value.comment != null) {
-      // add comment add on the line above
-      buffer.writeln();
-      buffer.writeln('\t/// ${value.comment}');
+    final generateAutodoc =
+        localeData.base && config.autodoc.enabled && value is LeafNode;
+    if (value.comment != null || generateAutodoc) {
+      if (value.comment != null) {
+        // add comment add on the line above
+        buffer.writeln();
+        buffer.writeln('\t/// ${value.comment}');
+      }
+
+      if (generateAutodoc) {
+        if (value.comment != null) {
+          buffer.writeln('\t///');
+        } else {
+          buffer.writeln();
+        }
+
+        for (int i = 0; i < config.autodoc.locales.length; i++) {
+          final locale = config.autodoc.locales[i];
+          final String? autodoc;
+          if (locale == AutodocConfig.base) {
+            autodoc = localeData.getAutodoc('', value as LeafNode);
+          } else {
+            final localeData = allTranslations.firstWhereOrNull(
+              (l) => l.locale.languageTag == locale,
+            );
+
+            if (localeData == null) {
+              throw 'Locale "$locale" not found in translations.';
+            }
+
+            autodoc = localeData.getAutodoc(value.path, null);
+          }
+
+          if (autodoc != null) {
+            if (i != 0) {
+              buffer.writeln('\t///');
+            }
+
+            buffer.writeln(
+              "\t/// ${locale == AutodocConfig.base ? localeData.locale.languageTag : locale}: '$autodoc'",
+            );
+            prevHasComment = true;
+          }
+        }
+      }
       prevHasComment = true;
     } else {
       if (prevHasComment) {