+++ slang/lib/src/runner/utils/read_analysis_file.dart
@@ -8,6 +8,7 @@ import 'package:slang/src/builder/utils/file_utils.dart';
 import 'package:slang/src/builder/utils/path_utils.dart';
 import 'package:slang/src/builder/utils/regex_utils.dart';
 import 'package:slang/src/runner/apply.dart';
+import 'package:slang/src/utils/log.dart' as log;
 
 const _supportedFiles = [FileType.json, FileType.yaml];
 
@@ -60,17 +61,17 @@ Map<I18nLocale, Map<String, dynamic>> readAnalysis({
     try {
       parsedContent = BaseDecoder.decodeWithFileType(fileType, content);
     } on FormatException catch (e) {
-      print('');
+      log.verbose('');
       throw 'File: ${file.path}\n$e';
     }
 
     if (locale != null) {
       _printReading(locale, file);
-      resultMap[locale] = {...parsedContent}..remove(INFO_KEY);
+      resultMap[locale] = {...parsedContent}..remove(infoKey);
     } else {
       // handle file containing multiple locales
       for (final entry in parsedContent.entries) {
-        if (entry.key.startsWith(INFO_KEY)) {
+        if (entry.key.startsWith(infoKey)) {
           continue;
         }
 
@@ -94,5 +95,5 @@ Map<I18nLocale, Map<String, dynamic>> readAnalysis({
 }
 
 void _printReading(I18nLocale locale, File file) {
-  print(' -> Reading <${locale.languageTag}> from ${file.path}');
+  log.verbose(' -> Reading <${locale.languageTag}> from ${file.path}');
 }