+++ slang_flutter/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 4.9.0
+
+- bump `slang` to `4.9.0`
+
+## 4.8.0
+
+- bump `slang` to `4.8.0`
+
 ## 4.7.0
 
 - bump `slang` to `4.7.0`